fix semaphore waiting bug on an unlocked object
This bug has been originally noticed by Romain Guilloteau.

When `wait()` is called on a semaphore, `pock_lockobj_lock()`
first checks if the object is unlocked (for a semaphore it means
that at least one resource is available). In this case, it marks
the object as locked (unavailable) and returns succesfully.

However, when the lock object is a semaphore, this logic is doubly
wrong as:

- The resource counter for the semaphore is not decreased although
  one resource is being consumed;
- The semaphore is locked although more resources can be available.

The first problem means that the resource count is wrong, and the
following scenario can happen with two tasks H and L:

- semaphore is initialized with 0 resources (counter = 0, blocked)
- H: semaphore is signaled (counter = 1, unblocked)
- L: semaphore is waited for => success (counter = 1, blocked)
- L: semaphore is waited for => block
- H: semaphore is signaled (counter = 2, unblocked)
- L: => task waiting for semaphore is unblocked (counter = 1, unblocked)
- L: semaphore is waited for => success (counter = 0, unblocked)

Here we can see that the semaphore has been signaled twice but
succesfully waited for thrice. This is due to the incorrect
*counter = 1, blocked* state which should never happen (the
semaphore cannot be blocked while having resources available).

The second problem means that the semaphore can be locked while
resources are still available (task H is high priority, task L
is low priority):

- semaphore is initialized with 0 resources (counter = 0, blocked)
- L: semaphore is waited for => block
- H: semaphore is signaled (counter = 1, unblocked)
- H: semaphore is signaled (counter = 2, unblocked) by high priority task
- L: => task waiting for semaphore is unblocked (counter = 1, blocked)
- L: semaphore is waited for => block

The same incorrect state *counter = 1, locked* happens, but here
L ends up blocked although a resource is available because the semaphore
is wrongly blocked.