fix two lock issues in lock objects
Issue 1
-------

Spurious wakeups can happen when waiting for a locked object, for
example we can be allowed to wake up but a thread with a higher priority
requests the object before we can get a chance to properly lock it.
This is accounted for by using a `while` loop around the wakeups.

However, while the spinlock on the lock object has been properly
released before going to sleep, in case we have to sleep again because
of a spurious wakeup the lock is released *again* without being
acquired.

Scenario with two tasks H (high priority) and L (low priority):

- H: lock(obj) and sleep for some time
- L: lock(obj) => put to sleep after abandoning the
     temporarily aquired spinlock on obj
- H: unlock(obj) [this unsleeps task L] and immediately lock(obj)
- L: notice that obj is still locked, abandon the spinlock on obj
     incorrectly since it was never reacquired

Issue 2
-------

The spinlock on the lock object is never reacquired after a wakeup.
We might start to lock the object and get interrupted in the middle
of the lock object manipulation, and since we do not hold the spinlock
some other higher priority thread might also modify the lock object,
leaving it in an inconsistent state.

Scenario:

- H: lock(obj) and sleep for some time
- L: lock(obj) => put to sleep after abandoning the
     temporarily acquired spinlock on obj
- H: unlock(obj) and sleep for a very short time
- L: start the process of locking obj without acquiring the
     spinlock
- H: lock(obj) => interrupt L and lock obj that L is
     in the process of locking
- L: resume the process of locking obj

Result: both H and L think they have acquired obj.

Fix
---

Relock the object when waking up, and release the spinlock after the
whole manipulation.