remove redundant boolean from lock object
THe `pok_lockobj_t` structure contains redundant data when it is used to
represent a semaphore. In this case, the following invariant must hold
at any time when the spinlock on the lock object is not held:

    obj->is_locked == (obj->current_value == 0)

Violating this invariant was the cause of the bug fixed in commit
122a082944d5d32029c2b4f289cfc14cce57f8ef.

While the `is_locked` information is not typically found in semaphore
structures because the `current_value` is needed and suffices to
represent the number of available resources in the semaphore, most of
the locking logic is common for all kinds of lock objects, hence the
use of `is_locked`.

By removing the `is_locked` field and using the `current_value` one
for all kind of lock objects, the redundant information is removed.
This simplifies the code and suppresses one possible source of
unsoundness if the invariant gets violated.