Clean every partition before building it
This way, the libpok is compiled with the needed configuration and
does not reuse a previously compiled one which might lead to
inconsistencies between definitions seen by the partitions and the one
compiled (or not) in libpok.

This will be removed when partitions use their own object directory to
hold libpok objects.